+++ ChangeLog.md
@@ -7,7 +7,8 @@
   - Add static bool VersionValid(const QString& szVersion);
   - Add static int VersionCompare(const QString &ver1, const QString &ver2);
 - Add set applicaiont font
-- Add audio and video settings
+- Add audio and video settings widget
+- Add CIconUtils
 - Modify CRecentMenu interface
 - CRecentMenu:
   - Add unlimited entries